fix: restore 19/19 smojol parse success after bind mount regression
Root cause: rm -rf output/rekt (old clean logic) changed the directory
inode, breaking the Docker bind mount ./output/rekt:/output. The
container's /output became unwritable, causing NoSuchFileException in
smojol's Files.createDirectories call for the new <prog>.cbl.report/
output structure introduced in v0.1.0-RC6.

Fixes:
- doctor.sh: add bind-mount health check in ensure_rekt_containers —
  probes /output/.write_probe and restarts cobol-rekt if stale, then
  aborts if still broken after restart
- doctor.sh: use find -mindepth 1 -delete (preserves inode) instead of
  rm -rf to prevent future regressions
- doctor.sh: per-file parse logs to output/rekt/<stem>.parse.log;
  deleted on success, kept with path hint on failure
- doctor.sh: run ID uses date +%Y%m%d%H%M (12-digit); McpChatWeb
  Program.cs reads it as long to avoid int32 overflow
- populator.py: add endswith('-data.json') to data-structures detection
  for smojol v0.1.0-RC6 naming (BDSM043.cbl-data.json)
- preprocess-for-rekt.sh: D-line fix (col 7 D → *DBG) and END-STRING
  insertion for programs that lacked explicit terminators

+++ McpChatWeb/Program.cs
@@ -1545,7 +1545,7 @@ app.MapGet("/api/graph/rekt/runs", async (CancellationToken cancellationToken) =
 		var runs = new List<object>();
 		await result.ForEachAsync(r => runs.Add(new
 		{
-			runId = r["runId"].As<int>(),
+			runId = r["runId"].As<long>(),
 			fileCount = r["fileCount"].As<int>(),
 			maxLines = r["maxLines"].As<int>()
 		}));
